Cache the dictionary mapping criteria/points to option IDs Reorganize assessment model tests.
